+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## v2.0.0 _17 June 2022_
 
+- fixed the caching not preventing requests being made
+- make the service resilient to Redis being unavailable
+- add a new simplified endpoint
 - Update Ruby
 - Update GEMS
 - Add lint step to build process